pyproject.toml
[project]
name = "iris-classifier"
version = "1.0.0"
description = "Iris flower classification model exposed via FastAPI REST API"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "joblib>=1.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
